% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellity-package.R
\docType{package}
\name{cellity-package}
\alias{cellity-package}
\title{Quality Control for Single-Cell RNA-seq Data}
\description{
\pkg{cellity} provides a support vector machine and PCA approaches to 
identifying and filtering low quality cells from single-cell RNA-seq datasets.
}

