% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extra_human_genes}
\alias{extra_human_genes}
\title{Additional human genes that are used in feature extraction}
\format{a list containing vectors of genes. Name indicates which GO category.}
\source{
Wellcome Trust Sanger Institute
}
\usage{
extra_human_genes
}
\value{
NULL, but makes available a list with  metadata
}
\description{
This list contains human genes that are used for feature extraction of biological features
}
\author{
Tomislav Ilicic & Davis McCarthy, 2015-03-05
}

