% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_qual_cells.R
\name{feature_generation}
\alias{feature_generation}
\title{Helper Function to create all features}
\usage{
feature_generation(counts_nm, read_metrics, GO_terms, extra_genes, organism)
}
\arguments{
\item{counts_nm}{Gene expression counts dataframe (genes x cells). Either 
normalised by library size or TPM values}

\item{read_metrics}{Dataframe with mapping statistics produced by python 
pipeline}

\item{GO_terms}{DataFrame with gene ontology term IDs, that will be used in 
feature extraction}

\item{extra_genes}{Additional genes used for feature extraction}

\item{organism}{The target organism to generate the features for}
}
\value{
Returns the entire set of features in a data.frame
}
\description{
Helper Function to create all features
}

