% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_info}
\alias{feature_info}
\title{Information which genes and GO categories should be included as features. Also defines which features are cell-type independent (common features)}
\format{a list with 2 elements (GO_terms,common_features).}
\source{
Wellcome Trust Sanger Institute
}
\usage{
feature_info
}
\value{
NULL, but makes available a list with  metadata
}
\description{
This list contains metadata information 
that is used to extract features from in the function extract_features
}
\author{
Tomislav Ilicic & Davis McCarthy, 2015-03-05
}

