% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_qual_cells.R
\name{info}
\alias{info}
\title{Internal function to print info string}
\usage{
info(string)
}
\arguments{
\item{string}{a string to print as an info message}
}
\description{
Internal function to print info string
}

