% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_qual_cells.R
\name{multiplot}
\alias{multiplot}
\title{Internal multiplot function to combine plots onto a grid}
\usage{
multiplot(..., plotlist = NULL, file, cols = 6, layout = NULL)
}
\arguments{
\item{...}{individual plots to combine into a single plot}

\item{plotlist}{a vector with names of plots to use in the plot}

\item{file}{string giving filename to which pdf of plots is to be saved}

\item{cols}{integer giving number of columns for the plot}

\item{layout}{matrix defining the layout for the plots}
}
\value{
a plot object
}
\description{
Internal multiplot function to combine plots onto a grid
}

