% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_qual_cells.R
\name{normalise_by_factor}
\alias{normalise_by_factor}
\title{Internal function to normalize by library size}
\usage{
normalise_by_factor(counts, norm_factor)
}
\arguments{
\item{counts}{matrix of counts}

\item{norm_factor}{vector of normalisation factors}
}
\value{
a matrix with normalized gene counts
}
\description{
Internal function to normalize by library size
}
\examples{
data(sample_counts)
data(sample_stats)
sample_counts_nm <- normalise_by_factor(sample_counts, colSums(sample_counts))
}

