% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{param_mES_common}
\alias{param_mES_common}
\title{Parameters used for SVM classification}
\format{a dataframe with 3 columns (gamma, cost, class.weights).}
\source{
Wellcome Trust Sanger Institute
}
\usage{
param_mES_common
}
\value{
NULL, but makes available a dataframe with parameters
}
\description{
This data frame has 3 columns:
gamma, cost, class.weights and is optimised for common features and our training data
}
\author{
Tomislav Ilicic & Davis McCarthy, 2015-03-05
}

