% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_qual_cells.R
\name{plot_pca}
\alias{plot_pca}
\title{Plots PCA of all features. Colors high and low quality cells based on outlier detection.}
\usage{
plot_pca(features, annot, pca, col, output_file)
}
\arguments{
\item{features}{Input dataset containing features (cell x features)}

\item{annot}{Matrix annotation of each cell}

\item{pca}{PCA of features}

\item{col}{color code indicating what color high and what low quality cells}

\item{output_file}{where plot is stored}
}
\value{
Plots of PCA
}
\description{
Plots PCA of all features. Colors high and low quality cells based on outlier detection.
}
\details{
This function plots PCA of all features + most informative features
}

