% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_counts}
\alias{sample_counts}
\title{Sample gene expression data containing 40 cells}
\format{a dataframe with genes x cells}
\source{
Wellcome Trust Sanger Institute
}
\usage{
sample_counts
}
\value{
NULL, but makes available a dataframe with raw read counts
}
\description{
This data frame contains genes (rows)
and cells (columns) showing raw read counts
}
\author{
Tomislav Ilicic & Davis McCarthy, 2015-03-05
}

