% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_stats}
\alias{sample_stats}
\title{Sample read statistics data containing 40 cells}
\format{a dataframe with cells x metrics}
\source{
Wellcome Trust Sanger Institute
}
\usage{
sample_stats
}
\value{
NULL, but makes available a dataframe with read statistics
}
\description{
This data frame contains read metrics (columns)
and cells (rows)
}
\author{
Tomislav Ilicic & Davis McCarthy, 2015-03-05
}

