% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_qual_cells.R
\name{simple_cap}
\alias{simple_cap}
\title{Converts all first letters to capital letters}
\usage{
simple_cap(x)
}
\arguments{
\item{x}{string}
}
\value{
a character vector in title case
}
\description{
Converts all first letters to capital letters
}

