% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_qual_cells.R
\name{sum_prop}
\alias{sum_prop}
\title{Sums up normalised values of genes to groups.}
\usage{
sum_prop(counts, genes_interest)
}
\arguments{
\item{counts}{Normalised gene expression count matrix}

\item{genes_interest}{dataframe of genes of interest to merge}
}
\value{
a vector of sums per group
}
\description{
Supports TPM and proportion of mapped reads.
}

