% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{training_mES_features}
\alias{training_mES_features}
\title{Original training dataset containing all and common features from the paper (training mES)}
\format{a list with 2 elements (all_features, common_features).}
\source{
Wellcome Trust Sanger Institute
}
\usage{
training_mES_features
}
\value{
NULL, but makes available a list with 2 dataframes
}
\description{
This list contains 2 dataframes
where each contains features per cell (cell X features) that can be used for classification.
}
\author{
Tomislav Ilicic & Davis McCarthy, 2015-03-05
}

