% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{training_mES_labels}
\alias{training_mES_labels}
\title{Original training dataset containing annotation of cells}
\format{a dataframe with 2 columns (cell_names, label).}
\source{
Wellcome Trust Sanger Institute
}
\usage{
training_mES_labels
}
\value{
NULL, but makes available a dataframe with cell annotations
}
\description{
This data frame has 2 columns:
First showing cell names, the second indicating if cell is of low (0) or high (1) quality
}
\author{
Tomislav Ilicic & Davis McCarthy, 2015-03-05
}

