% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_qual_cells.R
\name{uni.plot}
\alias{uni.plot}
\title{Internal function to detect outliers}
\usage{
uni.plot(x, symb = FALSE, quan = 1/2, alpha = 0.025, ...)
}
\arguments{
\item{x}{A matrix containing counts}
}
\description{
Internal function to detect outliers
}

