% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_qual_cells.R
\name{vote}
\alias{vote}
\title{Internal voting function to get final labels}
\usage{
vote(predictions)
}
\arguments{
\item{predicitions}{Predicted labels}
}
\description{
Internal voting function to get final labels
}

