% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_qual_cells.R
\name{assess_cell_quality_PCA}
\alias{assess_cell_quality_PCA}
\title{ASSESS CELL QUALITY USING PCA AND OUTLIER DETECTION}
\usage{
assess_cell_quality_PCA(features, file = "")
}
\arguments{
\item{features}{Input dataset containing features (cell x features)}

\item{file}{Output_file where plot is saved}
}
\value{
Returns a dataframe indicating which cell is low or high quality (0 
or 1 respectively)
}
\description{
ASSESS CELL QUALITY USING PCA AND OUTLIER DETECTION
}
\details{
This function applies PCA on features and uses outlier detection to
 determine which cells are low and which are high quality
}
\examples{
data(training_mES_features)
training_mES_features_all <- training_mES_features[[1]]
training_quality_PCA_allF <- assess_cell_quality_PCA(training_mES_features_all)
}

